Correções no Puppeteer

+++ backend/src/puppeteer/puppeteer.js
@@ -1,50 +1,67 @@
 const fs = require("fs");
 const s3 = require("../aws/s3");
 
-const randomUseragent = require("random-useragent");
 const puppeteer = require("puppeteer-extra");
 const StealthPlugin = require("puppeteer-extra-plugin-stealth");
 puppeteer.use(StealthPlugin());
 
 const filename = __filename.slice(__dirname.length + 1) + " -";
 
-// Constantes para serem utilziadas nos options dentro dos selects
+// Constantes para serem utilizadas nos options dentro dos selects
 const options = {
   dlog: "65776",
   correios: "63272",
-  PAC: "8899727307",
+  pac: "8899727307",
   sedex: "8899727306",
 };
 
 module.exports = {
   async removePopUp(page) {
-    page.setDefaultTimeout(15000);
-
     try {
-      await page.waitForXPath(
-        `//span[@class='menu-username hidden-xs hidden-sm']`,
-        {
-          visible: true,
-        }
-      );
+      console.log(filename, "Remoção de pop up iniciada.");
 
-      const popUpButton = await page.$x(
-        "//button[contains(text(), 'Não mostrar novamente')]"
-      );
+      await page.waitForXPath(`//button[contains(text(), 'Não mostrar novamente')]`, {
+        visible: true,
+        timeout: 15000,
+      });
 
+      const popUpButton = await page.$x("//button[contains(text(), 'Não mostrar novamente')]");
       await popUpButton[0].click();
 
-      console.log(filename, "Botão de remover Pop Up acionado.");
+      await page.waitForXPath(`//button[contains(text(), 'Não mostrar novamente')]`, {
+        hidden: true,
+        timeout: 15000,
+      });
+
+      console.log(filename, "Remoção de pop up concluído.");
     } catch (error) {
-      console.log(filename, `O popup já foi removido ou não foi encontrado`);
+      console.log(filename, `Não foi possível remover o pop up:`, error.message);
     }
-    page.setDefaultTimeout(100000);
+  },
+
+  async removerFiltroEmAberto(page) {
+    console.log(filename, "Removendo filtro de Em Aberto.");
+
+    await page.waitForXPath(`//span[text()='Em aberto']/following-sibling::span`, {
+      visible: true,
+      timeout: 15000,
+    });
+
+    console.log("clicando");
+    await page.waitForTimeout(1000);
+    const popUpButton = await page.$x(`//span[text()='Em aberto']/following-sibling::span`);
+    await popUpButton[0].click();
+
+    await page.waitForXPath(`//span[text()='Em aberto']/following-sibling::span`, {
+      hidden: true,
+      timeout: 15000,
+    });
+
+    console.log(filename, "Filtro de Em Aberto removido.");
   },
 
   async alterarTransportadora(pedido, metodo) {
-    // Cria uma nova instância do navegador
     try {
-      // Tenta iniciar uma nova instância do navegador
       const browser = await puppeteer.launch({
         headless: false,
         args: [`--window-size=1024,720`, "--no-sandbox", "--no-zygote"],
@@ -54,95 +71,23 @@ module.exports = {
         },
       });
 
-      // A instância foi iniciada com sucesso, prosseguir
       try {
-        // Tempo de inicio da operação
         let tempoInicio = new Date();
 
-        // Novo objeto de manipulação de página
         const page = await browser.newPage();
 
-        // Randomização
-        const userAgent = randomUseragent.getRandom();
+        await page.setUserAgent("Chrome/97.0.4692.71 (Windows NT 10.0; Win64; x64)");
 
-        const customUserAgent =
-          // "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/97.0.4692.71 Safari/537.36";
-          "Chrome/97.0.4692.71 (Windows NT 10.0; Win64; x64)";
-
-        await page.setUserAgent(customUserAgent);
-
-        // Pula verificações
-
-        //Skip images/styles/fonts loading for performance
-        // await page.setRequestInterception(true);
-        // page.on("request", (req) => {
-        //   if (
-        //     req.resourceType() == "stylesheet" ||
-        //     req.resourceType() == "font" ||
-        //     req.resourceType() == "image"
-        //   ) {
-        //     req.abort();
-        //   } else {
-        //     req.continue();
-        //   }
-        // });
-
-        await page.evaluateOnNewDocument(() => {
-          // Pass webdriver check
-          Object.defineProperty(navigator, "webdriver", {
-            get: () => false,
-          });
-        });
-
-        await page.evaluateOnNewDocument(() => {
-          // Pass chrome check
-          window.chrome = {
-            runtime: {},
-            // etc.
-          };
-        });
-
-        await page.evaluateOnNewDocument(() => {
-          //Pass notifications check
-          const originalQuery = window.navigator.permissions.query;
-          return (window.navigator.permissions.query = (parameters) =>
-            parameters.name === "notifications"
-              ? Promise.resolve({ state: Notification.permission })
-              : originalQuery(parameters));
-        });
-
-        await page.evaluateOnNewDocument(() => {
-          // Overwrite the `plugins` property to use a custom getter.
-          Object.defineProperty(navigator, "plugins", {
-            // This just needs to have `length > 0` for the current test,
-            // but we could mock the plugins too if necessary.
-            get: () => [1, 2, 3, 4, 5],
-          });
-        });
-
-        await page.evaluateOnNewDocument(() => {
-          // Overwrite the `languages` property to use a custom getter.
-          Object.defineProperty(navigator, "languages", {
-            get: () => ["en-US", "en"],
-          });
-        });
-
-        // Altera o timeout padrão da página
-        page.setDefaultTimeout(100000);
+        await page.setDefaultTimeout(100000);
 
         // Realiza Login
-        console.log(
-          filename,
-          `Pedido de Venda: ${pedido} -`,
-          "Realizando Login"
-        );
+        console.log(filename, `Pedido de Venda: ${pedido} -`, "Realizando Login");
         await page.goto("https://www.bling.com.br/login");
 
+        // Debug - Enviar imagem para o S3 do ambiente renderizado durante o login
         let port = process.env.PORT;
         if (port) {
-          //Rodando na AWS
-          // Debug
-          const screenshot = await page.screenshot({
+          await page.screenshot({
             path: "exports/puppeteer/screenshot.png",
           });
           const file = "exports/puppeteer/screenshot.png";
@@ -158,61 +103,34 @@ module.exports = {
         await page.click("button[name=enviar]");
 
         // Aguarda a homepage após login estar navegável
-
-        // Versão antiga:
-        // await page.waitForNavigation();
-
-        // Nova versão:
         // Para isso, aguardamos aparecer na tela o indicador do nome do usuario
-        await page.waitForXPath(
-          `//span[@class='menu-username hidden-xs hidden-sm']`,
-          {
-            visible: true,
-          }
-        );
+        await page.waitForXPath(`//span[@class='menu-username hidden-xs hidden-sm']`, {
+          visible: true,
+        });
 
         // Navegar para tela de pedidos
-        console.log(
-          filename,
-          `Pedido de Venda: ${pedido} -`,
-          "Navegando para tela de pedidos"
-        );
+        console.log(filename, `Pedido de Venda: ${pedido} -`, "Navegando para tela de pedidos");
         await page.goto("https://www.bling.com.br/vendas.php#list");
-        await page.waitForSelector("#pesquisa-mini", { visible: true });
 
         // No dia 28/06/2022 o Bling alterou a tela de pedidos e passou a exibir um popup
         // É necessário clicar no botão de "Não mostrar novamente" para não exibi-lo definitivamente
         await this.removePopUp(page);
-        //button[contains(text(), 'Não mostrar novamente')]
+
+        // Aguarda pela barra de pesquisa
+        await page.waitForSelector("#pesquisa-mini", { visible: true });
 
         // Buscar pelo pedido
-        console.log(
-          filename,
-          `Pedido de Venda: ${pedido} -`,
-          "Iniciando busca pelo pedido"
-        );
+        console.log(filename, `Pedido de Venda: ${pedido} -`, "Iniciando busca pelo pedido");
         await page.type("#pesquisa-mini", pedido);
         await page.click("#btn-mini-search");
 
         // Limpar tag de filtro em aberto
-        console.log(
-          filename,
-          `Pedido de Venda: ${pedido} -`,
-          "Limpando filtro em aberto"
-        );
-        await page.waitForSelector(".clear-tag", { visible: true });
-        await page.click(".clear-tag");
-        await page.waitForSelector(".clear-tag", { visible: true });
+        console.log(filename, `Pedido de Venda: ${pedido} -`, "Limpando filtro em aberto");
 
-        // Agora aguarda para verificar se a tabela de resultados apareceu
-        await page.waitForSelector("#datatable");
+        await this.removerFiltroEmAberto(page);
 
         // Adquirir Url do Pedido
-        console.log(
-          filename,
-          `Pedido de Venda: ${pedido} -`,
-          "Adquirindo URL do Pedido"
-        );
+        console.log(filename, `Pedido de Venda: ${pedido} -`, "Adquirindo URL do Pedido");
 
         //Aguarda pela visibilidade de um checkbox seguido do número do pedido que queremos alterar
         await page.waitForXPath(
@@ -225,15 +143,8 @@ module.exports = {
           `//span[contains(text(), '${pedido}')]/ancestor::tr//td[@class='checkbox-item']//div[@class='input-checkbox']//input`
         );
 
-        const idValue = await page.evaluate(
-          (x) => x.value,
-          idElementHandler[0]
-        );
-        console.log(
-          filename,
-          `Pedido de Venda: ${pedido} -`,
-          "ID para a URL: " + idValue
-        );
+        const idValue = await page.evaluate((x) => x.value, idElementHandler[0]);
+        console.log(filename, `Pedido de Venda: ${pedido} -`, "ID para a URL: " + idValue);
         const targetUrl = `https://www.bling.com.br/vendas.php#edit/${idValue}`;
 
         // Navegando para a página do pedido
@@ -246,25 +157,20 @@ module.exports = {
           "//select[@id='idConfUnidadeNegocio']//option[contains(text(), 'Matriz')]"
         );
 
-        // // Verifica se foi aberto o pedido correto
+        // Verifica se foi aberto o pedido correto
         await page.waitForXPath("//input[@name='numeroPedido']", {
           visible: true,
         });
 
         const numeroPedidoElement = await page.$("#numeroPedido");
 
-        const numeroPedido = await page.evaluate(
-          (x) => x.value,
-          numeroPedidoElement
-        );
+        const numeroPedido = await page.evaluate((x) => x.value, numeroPedidoElement);
         console.log(
           filename,
           `Pedido de Venda: ${pedido} -`,
           `Página do pedido de venda aberto:`,
           numeroPedido
         );
-
-        await this.removePopUp(page);
 
         if (numeroPedido != pedido) {
           console.log(
@@ -278,77 +184,53 @@ module.exports = {
 
         // Aguarda pelo select de integração logística estar visível e populado de opções
         await page.waitForSelector("#integracaoLogistica", { visible: true });
-        await page.waitForXPath(
-          "//select[@id='integracaoLogistica']//option[@value='-1']"
-        );
+        await page.waitForXPath("//select[@id='integracaoLogistica']//option[@value='-1']");
         await page.mainFrame().hover("#integracaoLogistica");
         // await this.delay(500);
 
         switch (metodo) {
-          case "dlog": {
+          case "dlog":
+          case "DLog": {
             // Caso 1 - DLOG
-            console.log(
-              filename,
-              `Pedido de Venda: ${pedido} -`,
-              "Alterando para: DLOG"
-            );
+            console.log(filename, `Pedido de Venda: ${pedido} -`, "Alterando para: DLOG");
             await page.select("select#integracaoLogistica", options.dlog);
             await this.delay(500);
             break;
           }
 
-          case "pac": {
+          case "pac":
+          case "PAC": {
             // Caso 2 - Correios PAC
-            console.log(
-              filename,
-              `Pedido de Venda: ${pedido} -`,
-              "Alterando para: Correios PAC"
-            );
+            console.log(filename, `Pedido de Venda: ${pedido} -`, "Alterando para: Correios PAC");
             await page.select("select#integracaoLogistica", options.correios);
             await this.delay(500);
 
             // Aguarda e posiciona no novo select que será renderizado
-            await page.waitForXPath(
-              "//select[@name='servicosLogistica[]']//option[@value='-1']"
-            );
-            const selectHoverPac = await page.waitForXPath(
-              "//select[@name='servicosLogistica[]']"
-            );
+            await page.waitForXPath("//select[@name='servicosLogistica[]']//option[@value='-1']");
+            const selectHoverPac = await page.waitForXPath("//select[@name='servicosLogistica[]']");
             await selectHoverPac.hover();
 
             // Seleciona o serviço
-            await page.select(
-              "select[name='servicosLogistica[]']",
-              options.PAC
-            );
+            await page.select("select[name='servicosLogistica[]']", options.pac);
             await this.delay(500);
             break;
           }
-
-          case "sedex": {
+          case "sedex":
+          case "SEDEX": {
             // Caso 2 - Correios sedex
-            console.log(
-              filename,
-              `Pedido de Venda: ${pedido} -`,
-              "Alterando para: Correios sedex"
-            );
+            console.log(filename, `Pedido de Venda: ${pedido} -`, "Alterando para: Correios sedex");
             await page.select("select#integracaoLogistica", options.correios);
             await this.delay(500);
 
             // Aguarda e posiciona no novo select que será renderizado
-            await page.waitForXPath(
-              "//select[@name='servicosLogistica[]']//option[@value='-1']"
-            );
+            await page.waitForXPath("//select[@name='servicosLogistica[]']//option[@value='-1']");
             const selectHoversedex = await page.waitForXPath(
               "//select[@name='servicosLogistica[]']"
             );
             await selectHoversedex.hover();
 
             // Seleciona o serviço
-            await page.select(
-              "select[name='servicosLogistica[]']",
-              options.sedex
-            );
+            await page.select("select[name='servicosLogistica[]']", options.sedex);
             await this.delay(500);
 
             break;
@@ -375,9 +257,7 @@ module.exports = {
 
         // Debug de tempo gasto na execução da alterçaão
         let tempoFinal = new Date();
-        let tempoGasto = new Date(tempoFinal - tempoInicio)
-          .toISOString()
-          .slice(11, -1);
+        let tempoGasto = new Date(tempoFinal - tempoInicio).toISOString().slice(11, -1);
         console.log(
           filename,
           `Pedido de Venda: ${pedido} -`,
@@ -406,8 +286,7 @@ module.exports = {
 
         // Inicia verificação para checar se o processo do browser foi finalizado corretamente
         // if (browser && browser.process() != null) browser.process().kill("SIGINT");
-        if (browser && browser.process() != null)
-          browser.process().kill("SIGKILL");
+        if (browser && browser.process() != null) browser.process().kill("SIGKILL");
       }
     } catch (error) {
       // O serviço não conseguiu iniciar uma nova instância do navegador
@@ -420,57 +299,6 @@ module.exports = {
       return false;
     }
   },
-
-  // async alterarTransportadoraMulti(pedidos) {
-  //   // Realizar Login apenas uma vez
-  //   // Percorrer cada um dos pedidos e alterar a transportadora
-  //   // Erros durante a alteração de um pedido não devem comprometer outros pedidos
-  //   // Finalizar o browser
-  //   // O Browser é finalizado apenas ao final da execução
-
-  //   try {
-  //     const browser = await puppeteer.launch({
-  //       headless: false,
-  //       args: [`--window-size=1024,720`, "--no-sandbox", "--no-zygote"],
-  //       defaultViewport: {
-  //         width: 1024 + Math.floor(Math.random() * 100),
-  //         height: 720 + Math.floor(Math.random() * 100),
-  //       },
-  //     });
-
-  //     // A instância foi iniciada com sucesso, prosseguir
-  //     try {
-  //       // Manusear a página aqui
-
-  //       // Tempo de inicio da operação
-  //       let tempoInicio = new Date();
-  //     } catch (error) {
-  //       // Erro durante o manuseio de um pedido em específico
-  //       console.log(
-  //         filename,
-  //         "Erro durante o manuseio da página:",
-  //         error.message
-  //       );
-
-  //       // Se o navegador não fechou, podemos iniciar o procedimento do próximo pedido
-  //       // ...
-  //     } finally {
-  //       console.log(
-  //         filename,
-  //         "Procedimento finalizado no Puppeteer, fechando navegador"
-  //       );
-  //       await browser.close();
-  //       if (browser && browser.process() != null)
-  //         browser.process().kill("SIGKILL");
-  //     }
-  //   } catch (error) {
-  //     console.log(
-  //       filename,
-  //       "Erro durante a abertura do navegador:",
-  //       error.message
-  //     );
-  //   }
-  // },
 
   async delay(tempo) {
     return new Promise((resolve) => {
